Делит на модули
